Jupyter Notebook
A machine learning project aimed at building a predictive model capable of estimating the rating of a movie based on various features, including genre, director, and actors. The primary objective is to harness regression techniques to tackle this problem and provide a tool that can help movie enthusiasts, filmmakers, and studios make informed decisions regarding movie production and marketing.